requirements.txt
Django==5.0.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
Pillow==10.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
requests==2.31.0
